backend/requirements/base.txt
python-dotenv
requests
python-firebase
openai
flask
firebase-admin
google-cloud-texttospeech
google-cloud-speech
fleep